js consolidation
